Refactor:refactoring initial code adding Register Controller
